some fixing
