+++ src/firebase/firebase.config.ts
@@ -12,4 +12,6 @@ export const firebaseConfig = {
 
 firebase.initializeApp(firebaseConfig);
 
+export const auth = firebase.auth();
+export const firestore = firebase.firestore();
 export default firebase;
