+++ src/sagas/index.sagas.tsx
@@ -1,4 +1,4 @@
-import sagas from './authSagas';
+import sagas from './auth/auth.sagas';
 import {all} from 'redux-saga/effects';
 
 export default function* rootSaga() {
